requirements.txt
fastapi[standard]==0.115.6
sqlmodel==0.0.22
pydantic-settings==2.7.1
asyncpg==0.30.0
alembic==1.14.0
mako==1.3.8
